server/requirements.txt
asynctest==0.12.2
aiohttp==3.3.2
aiounittest==1.1.0
pylint==1.8.3
pyserial==3.4
coverage==4.5.1
